Fix github issue # 991
Passes tests for issue 991. Dayjs and momentjs do not have identical
APIs. Original implementation of dayjs substituted all references to
"momentjs" with "dayjs". There is no dayjs.ISO_8601 constant. Dayjs
must be extended with a plugin to customize date parsing format. Dayjs
does not accept an array of date formats as an argument.