Refine typing for Column
_header, _key and width are dependent on defn and therefore may
be undefined.

Add comment for properties that are dependent on style. Since style will
be set to an empty object when defn is not provided, these properties
may be undefined as well.

Change references of Partial<Column> to Column since the undefined
properties are directly handled inside the Column type definition.