Replace unzipper with yauzl-promise
Unzipper had a transitive dependency on a package that does not have a
license. This means that every user is potentially violating the law in
using it. Personally, I think it's better to stick to following it, so
I've replaced the root dependency with a modern (and updated)
alternative.

However, yauzl-promise requires node v16 and above! I think there's some
bikeshedding to be done for whether this is worthwhile or not.