Issue: style.xml has [Object object] as formatCode
Somehow the excel file is in a state where exceljs does not handle
extracting the numFmt of a conditional format properly and instead has
it set to an object.

This results in the generate style.xml file to have a formatCode of
`[Object object]` inside which corrupts the file.

This fix is to check if the numFmt is of the appropriate string type and
if not, corrects it by extracting the formatCode before moving forward.